Fixed all compiler warnings
Added some comments
Cleaned up some files
Removed terminal_setcolor() until we figure out how we'll handle colors

+++ kernel/arch/i386/vga.h
@@ -3,6 +3,9 @@
 
 #include <stdint.h>
 
+// TODO: if we ever upgrade to c++, check to see if we can change this enum type
+// to uint_8 - there's no sense in declaring it as an int, but we are limited by
+// constraints of C language
 enum vga_color {
 	VGA_COLOR_BLACK = 0,
 	VGA_COLOR_BLUE = 1,
@@ -22,12 +25,16 @@ enum vga_color {
 	VGA_COLOR_WHITE = 15,
 };
 
+// Return the color code for printing to VGA given the foreground and background
+// colors
 static inline uint8_t vga_entry_color(enum vga_color fg, enum vga_color bg) {
-	return fg | bg << 4;
+	return (uint8_t)(fg | bg << 4); // TODO: define operator | for uint8_t
 }
 
+// Returns the code to print the character to VGA, given a character and color
+// information
 static inline uint16_t vga_entry(unsigned char uc, uint8_t color) {
-	return (uint16_t) uc | (uint16_t) color << 8;
+	return (uint16_t)(uc | color << 8); // TODO: define operator | for uint8_t
 }
 
 #endif
